Fix timestamp logging if LC_ALL or LC_TIME is set
This commit fixes an issue where some pre-, post- and uninstall
scripts output an error message on certain locales.

The affected command lines were prone to failure because they assumed
`LC_ALL` and `LC_TIME` to be unset (or set to an US-American locale).
To reproduce the issue, run:

$ bash -c 'export LC_ALL=; export LC_TIME=de_DE.UTF-8; date -j -f "%a %b %d %T %Z %Y" "`date`" "+%s"'

The reason why the original command uses that indirection is probably
due to miscommunication, see [1].

[1]: https://stackoverflow.com/q/38348963